@@ -2,27 +2,30 @@
   "name": "js_note_client",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.25",
-    "@fortawesome/free-solid-svg-icons": "^5.11.2",
-    "@fortawesome/react-fontawesome": "^0.1.7",
-    "axios": "^0.19.0",
-    "bulma": "^0.8.0",
-    "moment": "^2.24.0",
-    "node-sass": "^4.13.0",
-    "rbx": "^2.2.0",
-    "react": "^16.11.0",
-    "react-burger-menu": "^2.6.12",
-    "react-dom": "^16.11.0",
-    "react-quill": "^1.3.5",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.2.0"
+  "engines": {
+    "node": "18.16.1"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "BROWSER=none GENERATE_SOURCEMAP=false react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
+  },
+  "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "axios": "^1.4.0",
+    "bulma": "^0.9.4",
+    "moment": "^2.29.4",
+    "node-sass": "^9.0.0",
+    "rbx": "^2.2.0",
+    "react": "^16.14.0",
+    "react-burger-menu": "^3.0.9",
+    "react-dom": "^16.14.0",
+    "react-quill": "^2.0.0",
+    "react-router-dom": "^6.14.1",
+    "react-scripts": "^5.0.1"
   },
   "eslintConfig": {
     "extends": [
