@@ -1,74 +1,74 @@
-import React, { useState, useEffect } from "react";
-import "../../styles/notes.scss";
-import { push as Menu } from "react-burger-menu";
-import { Column, Button } from "rbx";
-import List from "../notes/list";
-import Editor from "../notes/editor";
-import Search from "../notes/search";
-import NotesService from "../../services/notes";
+import React, { useState, useEffect } from 'react'
+import '../../styles/notes.scss'
+import { push as Menu } from 'react-burger-menu'
+import { Column } from 'rbx'
+import List from '../notes/list'
+import Editor from '../notes/editor'
+import Search from '../notes/search'
+import NotesService from '../../services/notes'
 
 const Notes = (props) => {
-  const [notes, setNotes] = useState([]);
+  const [notes, setNotes] = useState([])
   const [current_note, setCurrentNote] = useState({
-    title: "",
-    body: "",
-    id: "",
-  });
+    title: '',
+    body: '',
+    id: ''
+  })
 
   async function fetchNotes() {
-    const response = await NotesService.index();
+    const response = await NotesService.index()
     if (response.data.length >= 1) {
-      setNotes(response.data.reverse());
-      setCurrentNote(response.data[0]);
+      setNotes(response.data.reverse())
+      setCurrentNote(response.data[0])
     } else {
-      setNotes([]);
+      setNotes([])
     }
   }
 
   const createNote = async () => {
-    await NotesService.create();
-    fetchNotes();
-  };
+    await NotesService.create()
+    fetchNotes()
+  }
 
   const deleteNote = async (note) => {
-    await NotesService.delete(note._id);
-    fetchNotes();
-  };
+    await NotesService.delete(note._id)
+    fetchNotes()
+  }
 
   const updateNote = async (oldNote, params) => {
-    const updatedNote = await NotesService.update(oldNote._id, params);
-    const index = notes.indexOf(oldNote);
-    const newNotes = notes;
-    newNotes[index] = updatedNote.data;
-    setNotes(newNotes);
-    setCurrentNote(updatedNote.data);
-  };
+    const updatedNote = await NotesService.update(oldNote._id, params)
+    const index = notes.indexOf(oldNote)
+    const newNotes = notes
+    newNotes[index] = updatedNote.data
+    setNotes(newNotes)
+    setCurrentNote(updatedNote.data)
+  }
 
   const searchNote = async (query) => {
-    const response = await NotesService.search(query);
-    setNotes(response.data);
-  };
+    const response = await NotesService.search(query)
+    setNotes(response.data)
+  }
 
   const selectNote = (id) => {
     const note = notes.find((note) => {
-      return note._id === id;
-    });
-    setCurrentNote(note);
-  };
+      return note._id === id
+    })
+    setCurrentNote(note)
+  }
 
   useEffect(() => {
-    fetchNotes();
-  }, []);
+    fetchNotes()
+  }, [])
 
   return (
     <>
       <Column.Group className="notes" id="notes">
         <Menu
-          pageWrapId={"notes-editor"}
+          pageWrapId={'notes-editor'}
           isOpen={props.isOpen}
           onStateChange={(state) => props.setIsOpen(state.isOpen)}
           disableAutoFocus
-          outerContainerId={"notes"}
+          outerContainerId={'notes'}
           customBurgerIcon={false}
           customCrossIcon={false}
         >
@@ -91,7 +91,7 @@ const Notes = (props) => {
         </Column>
       </Column.Group>
     </>
-  );
-};
+  )
+}
 
-export default Notes;
+export default Notes
