@@ -1,32 +1,32 @@
-import React, { useState, useEffect } from "react";
-import { Button, Field, Control, Input, Column, Title, Help, Label } from "rbx";
-import UsersService from "../../../services/users";
+import React, { useState, useEffect } from 'react'
+import { Button, Field, Control, Input, Column, Help, Label } from 'rbx'
+import UsersService from '../../../services/users'
 
 function UsersEditForm() {
-  const [email, setEmail] = useState("");
-  const [name, setName] = useState("");
-  const [status, setStatus] = useState(null);
+  const [email, setEmail] = useState('')
+  const [name, setName] = useState('')
+  const [status, setStatus] = useState(null)
 
   const initializeUser = async () => {
-    const user = await JSON.parse(localStorage.getItem("user"));
-    setName(user["name"]);
-    setEmail(user["email"]);
-  };
+    const user = await JSON.parse(localStorage.getItem('user'))
+    setName(user['name'])
+    setEmail(user['email'])
+  }
 
   useEffect(() => {
-    initializeUser();
-  }, []);
+    initializeUser()
+  }, [])
 
   const handleSubmit = async (evt) => {
-    evt.preventDefault();
+    evt.preventDefault()
 
     try {
-      await UsersService.update({ email: email, name: name });
-      setStatus("success");
+      await UsersService.update({ email: email, name: name })
+      setStatus('success')
     } catch (err) {
-      setStatus("error");
+      setStatus('error')
     }
-  };
+  }
 
   return (
     <>
@@ -36,7 +36,7 @@ function UsersEditForm() {
             <Label className="has-text-grey">Full Name</Label>
             <Input
               type="name"
-              value={name || ""}
+              value={name || ''}
               onChange={(e) => setName(e.target.value)}
               required
               name="name"
@@ -48,7 +48,7 @@ function UsersEditForm() {
             <Label className="has-text-grey">Email</Label>
             <Input
               type="email"
-              value={email || ""}
+              value={email || ''}
               onChange={(e) => setEmail(e.target.value)}
               required
               name="email"
@@ -67,13 +67,13 @@ function UsersEditForm() {
             </Column.Group>
           </Control>
         </Field>
-        {status == "error" && <Help color="danger">Problem in update</Help>}
-        {status == "success" && (
+        {status === 'error' && <Help color="danger">Problem in update</Help>}
+        {status === 'success' && (
           <Help color="primary">Updated with success</Help>
         )}
       </form>
     </>
-  );
+  )
 }
 
-export default UsersEditForm;
+export default UsersEditForm
