@@ -1,26 +1,26 @@
-import React, { Fragment, useState } from "react";
-import { Button, Field, Control, Input, Column, Title, Help, Label } from "rbx";
-import UsersService from "../../../services/users";
+import React, { Fragment, useState } from 'react'
+import { Button, Field, Control, Input, Column, Help, Label } from 'rbx'
+import UsersService from '../../../services/users'
 
 function UsersEditPasswordForm() {
-  const [password, setPassword] = useState("");
-  const [password_confirmation, setPasswordConfirmation] = useState("");
-  const [status, setStatus] = useState(null);
+  const [password, setPassword] = useState('')
+  const [password_confirmation, setPasswordConfirmation] = useState('')
+  const [status, setStatus] = useState(null)
 
   const handleSubmit = async (evt) => {
-    evt.preventDefault();
+    evt.preventDefault()
 
     if (password === password_confirmation) {
       try {
-        await UsersService.updatePassword({ password: password });
-        setStatus("success");
+        await UsersService.updatePassword({ password: password })
+        setStatus('success')
       } catch (err) {
-        setStatus("error");
+        setStatus('error')
       }
     } else {
-      setStatus("error_confirmation_password");
+      setStatus('error_confirmation_password')
     }
-  };
+  }
 
   return (
     <>
@@ -61,18 +61,18 @@ function UsersEditPasswordForm() {
             </Column.Group>
           </Control>
         </Field>
-        {status == "error_update" && (
+        {status === 'error_update' && (
           <Help color="danger">Problem in password update</Help>
         )}
-        {status == "error_confirmation_password" && (
+        {status === 'error_confirmation_password' && (
           <Help color="danger">Password don't match</Help>
         )}
-        {status == "success" && (
+        {status === 'success' && (
           <Help color="primary">Updated with success</Help>
         )}
       </form>
     </>
-  );
+  )
 }
 
-export default UsersEditPasswordForm;
+export default UsersEditPasswordForm
