@@ -1,5 +1,5 @@
-import axios from "axios";
+import axios from 'axios'
 
-const Api = axios.create({ baseURL: process.env.REACT_APP_BASE_API });
+const Api = axios.create({ baseURL: process.env.REACT_APP_BASE_API })
 
-export default Api;
+export default Api
