@@ -1,10 +1,10 @@
-const mongoose = require("mongoose");
-mongoose.Promise = global.Promise;
+const mongoose = require('mongoose')
+mongoose.Promise = global.Promise
 
-require("dotenv").config();
-const MONGO_URL = process.env.MONGO_URL;
+require('dotenv').config()
+const MONGO_URL = process.env.DATABASE_URL
 
 mongoose
   .connect(MONGO_URL)
-  .then(() => console.log("Connection succesful"))
-  .catch((err) => console.log(err));
+  .then(() => console.log('Connection succesful'))
+  .catch((err) => console.log(err))
